# Optimize the performance of `MessageSet#with`