@solnic 2019-11-05
Thank you. I'm curious now - did it actually cause performance issues for you?