@DNNX 2019-11-05
> did it actually cause performance issues for you?

~No, I was just hunting for Hacktoberfest swag.~  Not really, most of the time in my app is spent in the database.

`(a + b).uniq` pattern is just something on my radar. Every time I see it (or `a.map{}.flatten`, or `n.times.map{}`, or other suboptimal pattern), I trigger and send a pull request. A benchmark usually convinces people that the PR is well thought out and worth merging.

This trick doesn't work with Rails anymore, though :) They ask deep questions as to whether such micro-optimisation changes anything at the global level. Usually the answer is "no" and such pull request don't get merged.