@solnic 2019-11-05
@DNNX we accept most of the perf optimizations because even if something improves performance in less common situations it is still worth it (unless the optimization requires some really obscure tricks then more consideration is required)